% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java.R
\name{dr_java}
\alias{dr_java}
\alias{dr_h5reg}
\title{Check the local h5reg and java configuration (required for h5reg transforms)}
\usage{
dr_java()

dr_h5reg()
}
\description{
Check the local h5reg and java configuration (required for h5reg transforms)
}
\examples{
\donttest{
dr_h5reg()
}
}
