% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java.R
\name{dr_java}
\alias{dr_java}
\title{Check the local java configuration (required for h5reg transforms)}
\usage{
dr_java()
}
\description{
Check the local java configuration (required for h5reg transforms)
}
\examples{
\donttest{
dr_java()
}
}
