% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5reg.R
\name{h5reg-utils}
\alias{h5reg-utils}
\alias{read.h5reg.info}
\alias{is.h5reg}
\title{Read information about h5reg transformations from file}
\usage{
read.h5reg.info(x, ...)

is.h5reg(f = NULL, bytes = NULL)
}
\arguments{
\item{x}{path to a \code{.h5} file}

\item{...}{Currently ignored}

\item{f}{Path to a file on disk}

\item{bytes}{A set of bytes that are at least as big as the HDF5 magic value.}
}
\value{
a named list describing the contents of an \code{h5reg} file.

Logical indicating whether the file is an HDF5 encoded registration
  file ('code{h5reg})
}
\description{
\code{is.h5reg} checks if a file (or bytes in memory) looks like
  \code{h5reg} file.
}
\details{
if \code{bytes} is passed a raw byte array, then \code{is.h5reg} can
  do a very quick check to ensure that it is actually an HDF5 file, before
  inspect the contents of the file itself in more detail.
}
\examples{
\donttest{
h5f <- system.file('samples/complexdummyh5reg.h5', package = 'nat.h5reg')
read.h5reg.info(h5f)
is.h5reg(h5f)
}
}
\seealso{
\code{\link{h5reg}}
}
