% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5reg.R
\name{h5reg}
\alias{h5reg}
\title{Wrapper object for one or more HDF5 encoded transforms}
\usage{
h5reg(..., swap = NULL)
}
\arguments{
\item{...}{Path to one or more HDF5 encoded registrations.}

\item{swap}{When \code{TRUE} implies that the registration specifies the
direction required for transforming points. When \code{FALSE} implies the
inverse. The default implied by \code{swap=NULL} is \code{swap=TRUE} for
all registrations.}
}
\value{
A character vector with additional class \code{antsreg}.
}
\description{
Wrapping paths to one or more h5 registration files on disk in
  an \code{antsreg} object means that they can be used by
  \code{nat::\link{xform}} and inside \code{nat::\link{reglist}} objects.
}
\details{
h5 registrations contains both forward and inverse deformation
  fields along with the corresponding affine transforms wrapped in a single
  HDF5 file. They are therefore naively invertible just by picking out the
  correct registration field.

  For historical reasons (based on the convention of CMTK), when
  \code{swap=FALSE}, we assume we are trying to specify a registration in the
  direction required for image transformation. This is the opposite of the
  direction required for point transformation, which is the convention for h5
  registration files. We therefore set \code{swap=TRUE} when it is not
  specified.

  If multiple registrations are specified, they are given in order from
  floating to template brain, reading from left to right. See
  \code{nat::\link{xform}} and \code{nat::\link{reglist}} for further
  details.
}
\examples{
\dontrun{
# define forward and inverse registrations
inv=antsreg("JRC2018F_FAFB/JRC2018F_FAFB1InverseWarp_down.nii.gz",
    "JRC2018F_FAFB/JRC2018F_FAFB0GenericAffine.mat", swap=c(FALSE,TRUE))
fwd=antsreg("JRC2018F_FAFB/JRC2018F_FAFB0GenericAffine.mat",
    "JRC2018F_FAFB/JRC2018F_FAFB1Warp_down.nii.gz", swap=c(TRUE,TRUE))

# position of DA1 glomerulus in FAFB
da1glomr.fafb <- cbind(429316, 217924, 42960)
da1glomr.fafbum=da1glomr.fafb/1e3
# map position from FAFB to JRC2018
res <- xform(da1glomr.fafbum, inv)
# and back again
res2 <- xform(res, fwd)

# print out locations
da1glomr.fafbum
res
res2
}
}
\seealso{
Other h5reg: \code{\link{xformpoints.h5reg}}
}
\concept{h5reg}
