% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat.h5reg-package.R
\docType{package}
\name{nat.h5reg-package}
\alias{nat.h5reg}
\alias{nat.h5reg-package}
\title{nat.h5reg: Neuroanatomy Toolbox Support for Saalfeld/Bogovic HDF5 Registration Format}
\description{
Enables the Neuroanatomy Toolbox (nat) suite to carry out 3D
    coordinate transformation defined by points encoded in the HDF5
    registration format defined by Stephan Saalfeld and John Bogovic.
}
\section{h5 registration format}{
 The
  \href{https://github.com/saalfeldlab/template-building/wiki/Hdf5-Deformation-fields}{h5
  registration format} has been defined by John Bogovic and Stephan Saalfeld
  as an efficient way to distribute deformation fields (and accompanying
  affine transformations). This provides a standard way to distribute and
  apply such transforms in an HDF5 wrapper.

  Although deformation fields are the most general form of specifying a
  non-rigid registration, they can be much larger than a parametric
  deformation. h5 tools provide a substantial (lossy) compression of these
  fields at the expense of a very small (and usually irrelevant) loss of
  precision. They also provide the option to have one or more downsampled
  versions of the original field, again typically with relatively little
  impact on precision.

  This package supports use of `.h5` registration files through the
  \code{\link{h5reg}} class and associated functions.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/nat.h5reg}
  \item \url{https://natverse.github.io/nat.h5reg}
  \item Report bugs at \url{https://github.com/natverse/nat.h5reg/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

Authors:
\itemize{
  \item Saalfeld Lab (for transform-helpers java library) [copyright holder]
}

}
\keyword{internal}
