% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xform.R
\name{xformpoints.h5reg}
\alias{xformpoints.h5reg}
\title{Map points using h5 registrations}
\usage{
\method{xformpoints}{h5reg}(reg, points, ..., method = c("auto", "java", "rjava"))
}
\arguments{
\item{reg}{A \code{\link{h5reg}} object specifying one or more h5
registrations.}

\item{points}{An Nx3 matrix of 3D points to transform}

\item{...}{Additional arguments passed to java transform tool (currently
ignored).}

\item{method}{Whether to shell out to java executable (\code{method='java'})
or use the rJava package (\code{method='rjava'}). The default
(\code{method='auto'}) uses rJava if installed. See Java section for
implications of this choice.}
}
\value{
An Nx3 matrix of transformed points
}
\description{
Map points using h5 registrations
}
\details{
See \code{\link{h5reg}} for details of the ordering of
  registrations.
}
\section{Java}{
 By default \code{xformpoints.h5reg} starts a separate Java
  Virtual Machine (JVM) for every call. This has an overhead that might be in
  the 0.5s range, which quickly builds up. Alternatively, you can use
  \code{method='rjava'}, which depends on the 'rJava' package; since rJava
  can be a little fiddly to install, this is only a suggested dependency.
  Using rJava transparently starts a single JVM session from R and then
  reuses that, substantially speeding up the situation when you have many
  small objects to transform. It is also somewhat more efficient in terms of
  disk/CPU since points are retained in memory rather than written out as
  text files. However in my hands this has limited impact on the elapsed
  time.
}

\examples{
\donttest{
# basic usage
sampleh5path=system.file("samples/JRC2018F_FAFB_extrasmall.h5",
  package = 'nat.h5reg')
sampleh5reg=h5reg(sampleh5path)
xform(cbind(50,50,30), sampleh5reg)
xform(cbind(50,50,30), sampleh5reg, swap=TRUE)
}

\dontrun{
# specify a particular level for registration containing more than 1 level
# of detail
# NB not implemented in this sample
xform(cbind(50,50,30), sampleh5reg, level=0)
# choose faster, lower resolution registration
xform(cbind(50,50,30), sampleh5reg, level=2)
# print more detailed error messages when trying to debug
xform(cbind(50,50,30), sampleh5reg, level=0, stderr="")
}
}
\seealso{
\code{\link{h5reg}} for an example of point transformations using h5
  registrations and \code{\link{xform}} and \code{\link{xformpoints}} for
  details of how transformations are handled within the neuroanatomy toolbox
  suite.

Other h5reg: 
\code{\link{h5reg}()}
}
\concept{h5reg}
