% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5reg.R
\name{xformpoints.h5reg}
\alias{xformpoints.h5reg}
\title{Map points using h5 registrations}
\usage{
\method{xformpoints}{h5reg}(reg, points, ...)
}
\arguments{
\item{reg}{A \code{\link{h5reg}} object specifying one or more h5
registrations.}

\item{points}{An Nx3 matrix of 3D points to transform}

\item{...}{Additional arguments passed to java transform tool (currently
ignored).}
}
\value{
An Nx3 matrix of transformed points
}
\description{
Map points using h5 registrations
}
\details{
See \code{\link{h5reg}} for details of the ordering of
  registrations.
}
\examples{
\dontrun{
# basic usage
xform(cbind(50,50,30), h5reg('JRC2018F_FAFB.h5'))
xform(cbind(50,50,30), h5reg('JRC2018F_FAFB.h5', swap=TRUE))
# specify a particular level for registration containing more than 1 level
# of detail
xform(cbind(50,50,30), h5reg('JRC2018F_FAFB.h5'), level=0)
# print more detailed error messages when trying to debug
xform(cbind(50,50,30), h5reg('JRC2018F_FAFB.h5'), level=0, stderr="")
}
}
\seealso{
\code{\link{h5reg}} for an example of point transformations using h5
  registrations and \code{\link{xform}} and \code{\link{xformpoints}} for
  details of how transformations are handled within the neuroanatomy toolbox
  suite.

Other h5reg: \code{\link{h5reg}}
}
\concept{h5reg}
