% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5reg.R
\name{xformpoints.h5reg}
\alias{xformpoints.h5reg}
\title{Map points using h5 registrations}
\usage{
\method{xformpoints}{h5reg}(reg, points, ...)
}
\arguments{
\item{reg}{A \code{\link{h5reg}} object specifying one or more h5
registrations.}

\item{points}{An Nx3 matrix of 3D points to transform}

\item{...}{Additional arguments passed to java transform tool (currently
ignored).}
}
\value{
An Nx3 matrix of transformed points
}
\description{
Map points using h5 registrations
}
\details{
See \code{\link{h5reg}} for details of the ordering of
  registrations.
}
\seealso{
Other h5reg: \code{\link{h5reg}}
}
\concept{h5reg}
